requirements_lumax.txt
# Core API
fastapi
uvicorn
httpx
pydantic
redis
# GGUF Engine (llama.cpp)
llama-cpp-python
# Quantization & Performance
auto-gptq
optimum
bitsandbytes
# Audio & Senses
numpy
scipy
soundfile
# Personality & Biological Logic
python-dotenv
aiohttp
# Voice dependencies (torch comes in via torchvision; needed by body_interface EARS CUDA probe)
faster-whisper
piper-tts
librosa
instruct

compressed-tensors
qwen-vl-utils
torchvision
